requirements.txt
aiogram==3.*
python-dotenv==1.*
aiosqlite==0.20.0
aiohttp==3.*
aiofiles==25.1.0
pillow==10.*
pytesseract==0.3.*
pdf2docx==0.5.*
python-docx==1.*
python-pptx==1.*
